moving schemas to schemas file
