+++ rendermodules/point_match_optimization/optimization.html
@@ -10,7 +10,7 @@
       <li>{{ item[0] }}: {{ item[1] }}</li>
     {% endfor %}
     </ol>
-    <a href="{{ rs['tilepair_url'] }}">Tilepair URL</a>
+    <a href="{{ rs['tilepair_url'] }}" target="_blank">Tilepair URL</a>
     <br><br>
     <img src="{{ rs['match_img_filename'] }}", width="1600",height="800">
     <br>
